Fix critical bugs found in KinKal code review
- KKStrawMaterial: fix wallpath/wirepath typo in averagePathLengths
- KKStrawHit: fix copy constructor to copy dVar_/dDdT_ from rhs
- Chi2SHU: fix operator<< to use os parameter instead of std::cout
- DriftANNSHU: fix copy-paste error printing wrong config index
- KKFitUtilities/KKFit/KKStrawHitCluster: use double limits instead
  of float limits for min/max tracking
- KKFit: initialize mutable cache variables to zero
- KKShellXing: guard transitTime against division by zero
- KKFit: guard dsig calculation against zero docaVar

Agent-Logs-Url: https://github.com/Mu2e/Offline/sessions/c56c0d14-d58e-4d88-9e5a-7d64ba5efda6

+++ Mu2eKinKal/inc/KKStrawHitCluster.hh
@@ -143,7 +143,7 @@ namespace mu2e {
     // return time just before the first hit's time.  This insures hit clusters are updated before individual hits
     // This insures the weights subtracted correspond to the reference fit, and that any changes made to the
     // hits in the cluster get propagated to the residuals and weights before the next fit
-    double mintime(std::numeric_limits<float>::max());
+    double mintime(std::numeric_limits<double>::max());
     for(auto const& hit : hits_){
       mintime = std::min(hit->time(),mintime);
     }
